test(e2e/tui): reduce nudge suppression suite to wiring smoke tests
The 'continuation nudge fires when the orchestrator returns text-only
after a tool was called' case was flaky in CI (failed all 3 retries on
some runs, passed locally). It asserted that the continuation nudge
fires after a single user submit with the response sequence:

  1. assistant text + tool call -> tool executes
  2. assistant text-only post-tool

But the continuation nudge state machine suppresses the nudge while
toolsCalledSinceLastUserInput is true, so within a single user-input
cycle a post-tool text-only turn correctly does NOT fire the nudge
(pinned by continuation-nudge.test.ts:115-118). The E2E case only
ever passed because of incidental auto-compaction behavior that
injected a second user-input cycle, which is timing/context-window
dependent and unreliable on slower runners.

All the behavioral branches the file was testing (fresh-session
suppression, abort suppression, same-cycle vs new-cycle firing,
empty-turn after a tool) are exhaustively covered by:

  - src/extensions/orchestration/continuation-nudge.test.ts
  - src/extensions/prompt-construction/prompt-enrichment.test.ts

The E2E layer for this nudge only needs to prove the wiring is live
end-to-end. Reduce to two scenarios that do exactly that:

  1. Silent: text-only response in a fresh session -> no nudge phrase
     leaks into any subsequent provider request. Catches regressions
     where an unrelated path injects the nudge text.
  2. Fires: empty turn -> empty-turn nudge text reaches the next
     provider request. Asserts turn_end -> EmptyTurnNudge.evaluateTurn
     -> sendMessage(followUp) is alive.

Rename the file to nudge-wiring.test.ts to match the new focus.

Verified: 5 consecutive local runs pass, ~3.4s each, no retries.
